+++ admin-webapp/src/main/webapp/src/app/app.component.ts
@@ -5,7 +5,11 @@ import { User } from "./user/model/user.model";
 import { Router } from "@angular/router";
 import { LanguageStore } from "@sbac/rdw-reporting-common-ngx/i18n";
 
-const AdminPermissions = [ 'GROUP_WRITE', 'INSTRUCTIONAL_RESOURCE_WRITE' ];
+const AdminPermissions = [
+  'GROUP_WRITE',
+  'INSTRUCTIONAL_RESOURCE_WRITE',
+  'EMBARGO_WRITE'
+];
 
 @Component({
   selector: 'app-component',