+++ admin-webapp/src/main/webapp/src/app/embargo/embargo-scope.enum.ts
@@ -1,0 +1,9 @@
+/**
+ * The type of test results governed by the embargo setting
+ */
+export enum EmbargoScope {
+
+  Individual = 'Individual',
+  Aggregate = 'Aggregate'
+
+}