+++ reporting-service/src/test/resources/integration-test-data.sql
@@ -4,12 +4,12 @@ insert into school_group (id, natural_id, name) values
 insert into district_group (id, natural_id, name) values
   (-10, 'districtGroup1', 'districtGroup1');
 
-insert into district (id, natural_id, name, embargo_enabled) values
-  (-10, 'districtNat1', 'district1', 0),
-  (-20, 'districtNat2', 'district2', 0),
-  (-30, 'districtNat3', 'district3', 0);
+insert into district (id, natural_id, name) values
+  (-10, 'districtNat1', 'district1'),
+  (-20, 'districtNat2', 'district2'),
+  (-30, 'districtNat3', 'district3');
 
-insert into school (id, district_id, natural_id, name, district_embargo_enabled, update_import_id, updated, migrate_id, school_group_id, district_group_id, external_id) VALUES
+insert into school (id, district_id, natural_id, name, embargo_enabled, update_import_id, updated, migrate_id, school_group_id, district_group_id, external_id) VALUES
   (-10, -10, 'schoolNat1', 'school1', 0, -1, '1997-07-18 20:14:34.000000', -1, null, null, 'externalId1'),
   (-20, -10, 'schoolNat2', 'school2', 0, -1, '1997-07-18 20:14:34.000000', -1, -10, null, 'externalId2'),
   (-30, -20, 'schoolNat3', 'school3', 0, -1, '1997-07-18 20:14:34.000000', -1, null, -10, 'externalId3'),