+++ webapp/src/main/webapp/src/app/aggregate-report/aggregate-reports.component.ts
@@ -1,0 +1,9 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: 'aggregate-reports',
+  templateUrl: './aggregate-reports.component.html',
+})
+export class AggregateReportsComponent {
+
+}