+++ webapp/src/main/webapp/src/app/aggregate-report/results/aggregate-reports-results.component.ts
@@ -1,0 +1,9 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: 'aggregate-reports-results',
+  templateUrl: './aggregate-reports-results.component.html',
+})
+export class AggregateReportsResultsComponent {
+
+}