+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added zuul service gateway to webapp.
     * Webapp now proxies requests to services and authenticates using JWT
     * Additional zuul and jwt properties are now required in the configuration
+    * Additional service pods required in deployment, service/ELB/CNAME changes required to differentiate webapp from reporting-service pods
 
 #### 1.0.2 (Admin) - 2017-12-05